requirements.txt
asgiref==3.10.0
attrs==25.4.0
babel==2.17.0
colorama==0.4.6
Django==5.2.7
django-background-tasks==1.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docxcompose==1.4.0
docxtpl==0.20.1
drf-yasg==1.21.11
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
packaging==25.0
psycopg2==2.9.11
PyJWT==2.10.1
python-docx==1.2.0
python-dotenv==1.2.1
python-environ==0.4.54
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
